Initial Commit
Current docker image is based on alpine:3.15.  The docker image
downloads the latest-gcc-arm-none-eabi toolchain from the official arm
repo (uses a script to parse webdata).  It then adds the path to the
toolchain's binaries to the $PATH environment variable inside the docker
container.  This is achieved via the container_entrypoint.sh script
which runs on container run.