Replaced Base Image with Ubuntu
There was a nasty warning with alpine about not being able to obtain the
version information from libstdc++.so.6 even with the libstdc++ package
installed.  It has to do with the fact that alpine uses musl libc as
default, and even the frovlad/alpine-glibc does not address this
"incompatabillity."  Given that the toolchain is so large, the base
image being somewhat larger is acceptable;  yields from image size
optimization will be larger if focused on trimming the toolchain.

Ubuntu is chosen as replacement given its popularity and inclusion of
glibc.